requirements.txt
Django>=4.0.4,<4.1
djangorestframework>=3.13.1,<3.14
psycopg2>=2.9.3,<2.10
django-widget-tweaks==1.4.12
pillow==9.3.0
djangorestframework>=3.13.1,<3.14